+++ main.py
@@ -3,7 +3,6 @@ from urllib.parse import parse_qs, urlparse
 from typing import List, Optional, Tuple, Union
 
 from dotenv import load_dotenv
-from supabase import create_client, Client
 import logging
 import os
 import polars as pl
@@ -12,7 +11,6 @@ from datetime import datetime
 import yt_dlp
 from fasthtml.common import *
 from monsterui.all import *
-from tenacity import retry, stop_after_attempt, wait_exponential
 
 from utils import (calculate_engagement_rate, format_duration, format_number,
                    process_numeric_column)
@@ -22,73 +20,13 @@ from constants import (PLAYLIST_STEPS_CONFIG, FLEX_COL, FLEX_CENTER,
                        FLEX_BETWEEN, GAP_2, GAP_4, SECTION_BASE, CARD_BASE,
                        HEADER_CARD, FORM_CARD, NEWSLETTER_CARD)
 from validators import YoutubePlaylist, YoutubePlaylistValidator
+from db import setup_logging, init_supabase, supabase_client
 
 # Get logger instance
 logger = logging.getLogger(__name__)
 
-
-def setup_logging():
-    """Configure logging for the application.
-    
-    This function should be called at application startup.
-    It configures the logging format and level.
-    """
-    logging.basicConfig(
-        level=logging.INFO,
-        format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
-        datefmt='%Y-%m-%d %H:%M:%S')
-
-
 # Load environment variables
 load_dotenv()
-
-# Global Supabase client
-supabase_client: Optional[Client] = None
-
-
-@retry(stop=stop_after_attempt(3),
-       wait=wait_exponential(multiplier=1, min=4, max=10),
-       reraise=True)
-def init_supabase() -> Optional[Client]:
-    """Initialize Supabase client with retry logic and proper error handling.
-    
-    Returns:
-        Optional[Client]: Supabase client if initialization succeeds, None otherwise
-        
-    Note:
-        - Retries up to 3 times with exponential backoff
-        - Returns None instead of raising exceptions
-        - Logs errors for debugging
-    """
-    global supabase_client
-
-    # Return existing client if already initialized
-    if supabase_client is not None:
-        return supabase_client
-
-    try:
-        url: str = os.environ.get("NEXT_PUBLIC_SUPABASE_URL", "")
-        key: str = os.environ.get("NEXT_PUBLIC_SUPABASE_ANON_KEY", "")
-
-        if not url or not key:
-            logger.warning(
-                "Missing Supabase environment variables - running without Supabase"
-            )
-            return None
-
-        client = create_client(url, key)
-
-        # Test the connection
-        client.auth.get_session()
-
-        supabase_client = client
-        logger.info("Supabase client initialized successfully")
-        return client
-
-    except Exception as e:
-        logger.error(f"Failed to initialize Supabase client: {str(e)}")
-        return None
-
 
 # CSS Classes
 CARD_BASE_CLS = CARD_BASE
@@ -134,15 +72,13 @@ def init_app():
 
     # Initialize Supabase client
     try:
-        global supabase_client
-        supabase_client = init_supabase()
+        init_supabase()
         if supabase_client is None:
             logger.warning("Running without Supabase integration")
     except Exception as e:
         logger.error(
             f"Unexpected error during Supabase initialization: {str(e)}")
         # Continue running without Supabase
-        supabase_client = None
 
 
 # Initialize the application
